application-source/requirements.txt
django
google-genai
gunicorn
python-dotenv
uvicorn
deepgram-sdk
instaloader
yt-dlp
django-q2
django-picklefield
ffmpeg-python